Bulk update. Part of changelog follows:
eMule 0.50b (unofficial)
-----------------------
- Jul, 28. 2016       -
-----------------------
.: The current build environment is Visual Studio 2015 C++ (VS2013 might be good enough too).
.: ATL Server library is required since that code is not a part of Visual Studio or SDK any more.
.: Updated libraries to the latest versions. Currently: CxImage 7.02 lite (sourceforge), miniupnpc 2.0, crypto++ 5.6.3, libpng 1.6.23, ResizableLib (github), id3lib 3.9.0 (sourceforge).
.: Renamed directory "crypto51" to "cryptopp" because any version of crypto++ library starting from 5.1 might be used.
.: Removed miniupnpc library source code from srchybrid directory as its latest versions can be used "as is"
.: Note: Updated id3lib dropped Xing VBR support.
.: Manifest is necessary for getting correct Windows versions above 8.0
.: Changed default: Standard slot is 25 kB/s instead of 50 kB/s in official.
.: Changed default: Overlapped sockets are now enabled by default (should not cause crashes)
.: Changed default: AdjustNTFSDaylightFileTime is disabled by default. With this setting enabled all shared files were rehashed on DST on/off
.: Debug builds would not crash when loading mp3 tags
.: Debug build: disabled several asserts or changed conditions to disable pointlessly frequent interrupts.
.: Fixed issues with 'Minimize to tray' feature (program window was getting stuck in either minimized or restored state)
.: New feature: Single click restore from system tray when MiniMule is disabled.
.: Added more internal state flags to disable certain actions and avoid crashes while eMule is initializing or exiting.
.: Added exception handling when loading IP filter for closing open files on error.
.: Added/modified exception handling in several other potential trouble spots.
.: KAD page: implemented simpler progress indicator of initial bootstrap with special nodes file (remove failed nodes from Contacts list).
.: KAD page: fixed Contacts/Search Details button width issue (the button shrinks when Windows GUI settings changed).
.: KAD page: fixed wrong text (search term) display for Contacts/Search Details button.
.: Shared files page: fixed shared parts bar width issue (width changed when scrolling list right and left).
.: Fixed stored searches bug (multiple records for the same file from different servers saved incorrect number of sources).
.: IRC: text colours in channel window mimic mIRC's defaults.
.: IRC: added parsing of time stamp and some other minor changes.
.: Downloads list: Sorting by Progress uses exact ratio instead of rounded percentage values.
.: Downloads list: 'Next re-ask time' in tooltip would not be shown while peer is uploading.
.: Downloads list: Last seen complete time in Details dialog is updated whenever peers report having all parts.
.: Keep statistics values for files when purging old records in known2_64.met
.: Minor graphical changes: checkboxes placement in First Runtime Wizard; smaller font size in About... dialog for Debug build's longer program title.
.: Yes/No buttons in exit confirmation dialog replaced with Ok/Cancel in order to enable Esc key as Cancel button.
.: KAD bootstrap nodes build: fixed crash when loading nodes file
.: Implemented some small changes posted in the forum (tHeWiZaRdOfDoS, Enig123, DolphinX, Eulero, ...)
.: A few tiny fixes; too small and too numerous to be mentioned here; a few typos were fixed too (in comments and in identifiers as well)
.: The last but not least: lots of small code changes, optimizations and fixes suggested by static code analyzers.
.: Special thanks to Ergol for stress and endurance testing

*** That part should have been written by official developers (maybe later). Something like this: ***

eMule 0.50b beta
-----------------------
- Apr, 5. 2010+       -
-----------------------
.: Added overlapped sockets support.
.: Dropped KAD 1.0 support and removed related code.
.: Added display of KAD connection progress when using special bootstrap file.

+++ CatDialog.cpp
@@ -17,8 +17,6 @@
 #include "stdafx.h"
 #include "emule.h"
 #include "CustomAutoComplete.h"
-#include "Preferences.h"
-#include "otherfunctions.h"
 #include "SharedFileList.h"
 #include "emuledlg.h"
 #include "TransferDlg.h"
@@ -103,8 +101,8 @@ void CCatDialog::UpdateData()
 	CheckDlgButton(IDC_REGEXPR,m_myCat->ac_regexpeval);
 
 	newcolor = m_myCat->color;
-	m_ctlColor.SetColor(m_myCat->color == -1 ? m_ctlColor.GetDefaultColor() : m_myCat->color);
-	
+	m_ctlColor.SetColor(m_myCat->color == (DWORD)-1 ? m_ctlColor.GetDefaultColor() : m_myCat->color);
+
 	GetDlgItem(IDC_AUTOCATEXT)->SetWindowText(m_myCat->autocat);
 
 	m_prio.SetCurSel(m_myCat->prio);
@@ -134,7 +132,7 @@ void CCatDialog::Localize()
 
 	SetWindowText(GetResString(IDS_EDITCAT));
 
-	SetDlgItemText(IDC_STATIC_REGEXP,GetResString(IDS_STATIC_REGEXP));	
+	SetDlgItemText(IDC_STATIC_REGEXP,GetResString(IDS_STATIC_REGEXP));
 
 	m_prio.ResetContent();
 	m_prio.AddString(GetResString(IDS_PRIOLOW));
@@ -144,7 +142,7 @@ void CCatDialog::Localize()
 }
 
 void CCatDialog::OnBnClickedBrowse()
-{	
+{
 	TCHAR buffer[MAX_PATH] = {0};
 	GetDlgItemText(IDC_INCOMING, buffer, _countof(buffer));
 	if (SelectDir(GetSafeHwnd(), buffer,GetResString(IDS_SELECT_INCOMINGDIR)))
@@ -156,10 +154,10 @@ void CCatDialog::OnBnClickedOk()
 	CString oldpath = m_myCat->strIncomingPath;
 	if (GetDlgItem(IDC_TITLE)->GetWindowTextLength()>0)
 		GetDlgItem(IDC_TITLE)->GetWindowText(m_myCat->strTitle);
-	
+
 	if (GetDlgItem(IDC_INCOMING)->GetWindowTextLength()>2)
 		GetDlgItem(IDC_INCOMING)->GetWindowText(m_myCat->strIncomingPath);
-	
+
 	GetDlgItem(IDC_COMMENT)->GetWindowText(m_myCat->strComment);
 
 	m_myCat->ac_regexpeval= IsDlgButtonChecked(IDC_REGEXPR)>0;
